extracted annotation
