+++ autofit/mapper/__init__.py
@@ -1,4 +1,5 @@
-from autofit.mapper.prior_model.prior_model import PriorModel, AnnotationPriorModel
+from autofit.mapper.prior_model.prior_model import PriorModel
+from autofit.mapper.prior_model.annotation import AnnotationPriorModel
 from autofit.mapper.prior_model.collection_prior_model import CollectionPriorModel
 from autofit.mapper.prior_model.abstract_prior_model import AbstractPriorModel
 from .model_mapper import ModelMapper
